finished simple test suite for cc version

+++ cc/testing/TestCases.hh
@@ -1,6 +1,8 @@
 #ifndef _TEST_CASES_HH_
 #define _TEST_CASES_HH_
 
+#include <cstddef>
+
 #include "../inc/extint.hh"
 
 class TestCases final {
@@ -13,8 +15,8 @@ class TestCases final {
 
   private:
     uint128_t* test_cases_;
-    int test_size_;
-    int test_case_;
+    size_t test_size_;
+    size_t test_case_;
 };
 
 #endif // _TEST_CASES_HH_
